backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
argon2-cffi==23.1.0
Pillow==11.1.0
pyotp==2.9.0
psycopg2-binary==2.9.10
boto3==1.37.11
razorpay==1.4.2
google-auth==2.38.0
gunicorn==23.0.0
dj-database-url==2.3.0
setuptools>=69.0.0
